move assert to test_regression
